style(snd,hda): satisfy rustfmt and clippy on the audio path
The `workspace` CI job runs `cargo fmt --all -- --check` and clippy with
`-D warnings`; two findings belonged to this audio work:

- rustfmt disagreed with two hand-wrapped closures in snd.rs.
- clippy's `precedence` lint on the HDMI infoframe checksum, where
  `0x100 - (sum & 0xff) as u16 as u32 & 0xff` mixed `-` and `&` and
  relied on the reader knowing which binds tighter. Same value, now
  parenthesised explicitly and with the casts that did nothing removed,
  plus a line saying what the checksum is for.

`cargo fmt -p linux-object` also normalised one line in loader/abi.rs
that arrived unformatted with 8e65116; it is pure formatting and the
package cannot be checked without it.

+++ drivers/src/audio/hda.rs
@@ -607,7 +607,8 @@ impl HdaInner {
         frame[4] = channels - 1; // CC, coding type "refer to stream"
                                  // frame[8] = CA (0 = FL/FR), rest zero.
         let sum: u32 = frame.iter().map(|&b| b as u32).sum();
-        frame[3] = (0x100 - (sum & 0xff) as u16 as u32 & 0xff) as u8;
+        // Checksum: the bytes of the frame plus this one must sum to 0 mod 256.
+        frame[3] = ((0x100 - (sum & 0xff)) & 0xff) as u8;
 
         let _ = self.cmd(pin, VERB_SET_DIP_INDEX, 0);
         for &b in &frame {
